+++ CHANGES.md
@@ -11,6 +11,7 @@
    - Article storage switch console command supports custom storage backends
  - Implemented bug fixes and changes present in Znuny version 6.0.36:
    - Filter column settings only include unique column names
+   - A warning message of an empty article subject is now shown only when the subject is actually empty
 
 # 6.0.34 2022-05-06 - Security release
  - Fixed the security issue described in OTRS AG security advisory [OSA-2022-03](https://otrs.com/release-notes/otrs-security-advisory-2022-03/) and Znuny security advisory [ZSA-2022-02](https://www.znuny.org/en/advisories/zsa-2022-02)